# docs/openapi.yaml
openapi: 3.1.0
info:
  title: smalltalk API
  version: 0.2.0
  summary: Room chat joined by invite link or QR. No app install, no sign-up.
  description: |
    REST API and WebSocket contract of the smalltalk server (`backend/`).

    **Conventions**
    - JSON fields are `snake_case`; times are RFC 3339 in UTC, e.g. `2026-09-26T06:00:00Z`.
    - Every error uses the same envelope: `{"error": {"code": "invite_expired", "message": "..."}}`.
      Clients should branch on `code`, never on `message`.
    - Authentication is a short-lived JWT (12 h) bound to **one room and one member**. It is returned
      by `POST /rooms`, `POST /join` and `POST /rooms/{id}/owner-session` and sent as
      `Authorization: Bearer <jwt>`. A JWT for room A is rejected on every route of room B.
    - Membership is re-checked on every request: kicked or banned members get `403` even with a
      JWT that has not expired yet.

    **Realtime** traffic goes over the WebSocket at `GET /ws`; see that operation and the
    `ClientFrame` / `ServerFrame` schemas.

    **Rate limits**: `POST /join` 10/min per IP; sending 5 messages/s per member;
    image uploads 10/min per member (bursts of 5). Exceeding one returns `429 rate_limited`.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8000/api
    description: Docker Compose (through Nginx, which strips the /api prefix)
  - url: http://localhost:8080
...
security:
  - roomJWT: []
paths:
...
  /rooms:
    post:
      tags: [rooms]
      operationId: createRoom
      summary: Create a room
      description: |
        Creates a room and its host (role `owner`). Returns the host's JWT plus an `owner_token`
        that is shown only once; store it to get a new JWT later with
        `POST /rooms/{id}/owner-session`. No invite is created; call `POST /rooms/{id}/invites`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
            example:
              name: ทีมงานอีเวนต์
              display_name: กาย
      responses:
        "201":
          description: Room created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedRoom"
        "400":
          $ref: "#/components/responses/InvalidInput"
  /rooms/{id}/owner-session:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    post:
      tags: [rooms]
      operationId: createOwnerSession
      summary: Get a new host JWT with the owner token
      description: Lets the host come back after their JWT expired, from any device holding the owner token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [owner_token]
              properties:
                owner_token:
                  type: string
      responses:
        "200":
          description: A fresh session for the room's owner member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /rooms/{id}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    get:
      tags: [rooms]
      operationId: getRoom
      summary: Get room details
      responses:
        "200":
          description: The room.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Removed"
        "404":
          $ref: "#/components/responses/NotFound"
  /rooms/{id}/close:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    post:
      tags: [rooms]
      operationId: closeRoom
      summary: Close the room (host only)
      description: |
        Sets the room to `closed`, sends `room_closed` to every open WebSocket and closes them with
        code 4004. Members can still read history; nobody can join, send or upload any more.
      responses:
        "204":
          description: Room closed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenOrRemoved"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /rooms/{id}/members:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    get:
      tags: [members]
      operationId: listMembers
      summary: List active members with presence
      description: Kicked and banned members are excluded. `online` is true while the member has at least one open WebSocket.
      responses:
        "200":
          description: Members, oldest first.
          content:
            application/json:
              schema:
                type: object
                required: [members]
                properties:
                  members:
                    type: array
                    items:
                      $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Removed"
        "404":
          $ref: "#/components/responses/NotFound"
  /rooms/{id}/members/{mid}/kick:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - name: mid
        in: path
        required: true
        description: Member to remove.
        schema:
          type: string
          format: uuid
    post:
      tags: [members]
      operationId: kickMember
      summary: Kick or ban a member (host only)
      description: |
        Removes the member and closes their WebSockets immediately (frame `kicked`, close code 4003).
        A kicked person may come back through a valid invite as a **new** member; a banned member
        record can never be used again. Bans apply to the member, not the person, because there are
        no accounts: revoke and reissue invites to keep someone out. The host cannot be removed.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ban:
                  type: boolean
                  default: false
      responses:
        "204":
          description: Member removed.
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenOrRemoved"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    roomJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Room-bound JWT from `POST /rooms`, `POST /join` or `POST /rooms/{id}/owner-session`. Valid for 12 hours.
  parameters:
    RoomId:
      name: id
      in: path
      required: true
      description: Room id. Must match the room of the JWT.
      schema:
        type: string
        format: uuid
  responses:
    InvalidInput:
      description: "`invalid_input` (or `message_too_long`): the request is malformed; `message` explains why."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: invalid_input, message: display_name is required }
    Unauthorized:
      description: "`unauthorized`: missing, expired or invalid JWT, or a JWT for another room."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: unauthorized, message: missing or invalid credentials }
    Removed:
      description: "`kicked` or `banned`: the member was removed from the room."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: banned, message: you are banned from this room }
    ForbiddenOrRemoved:
      description: "`forbidden` (not the host), `kicked` or `banned`."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: forbidden, message: not allowed }
    NotFound:
      description: "`not_found`: the room, member, invite or image does not exist."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: not_found, message: not found }
...
  schemas:
    ErrorCode:
      type: string
      enum:
        - not_found
        - unauthorized
        - forbidden
        - invalid_input
        - invite_invalid
        - invite_expired
        - invite_exhausted
        - room_closed
        - room_full
        - rate_limited
        - banned
        - kicked
        - message_too_long
        - image_too_large
        - image_unsupported
        - internal
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
              description: Human-readable English detail; not for display logic.
...
    DisplayName:
      type: string
      minLength: 1
      maxLength: 32
      description: Trimmed; inner whitespace collapsed. Counted in characters, so Thai text is fine.
    Role:
      type: string
      enum: [owner, member]
    CreateRoomRequest:
      type: object
      required: [name, display_name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        display_name:
          $ref: "#/components/schemas/DisplayName"
    Session:
      type: object
      required: [room_id, member_id, role, jwt, expires_at]
      properties:
        room_id:
          type: string
          format: uuid
        member_id:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/Role"
        jwt:
          type: string
        expires_at:
          type: string
          format: date-time
    CreatedRoom:
      allOf:
        - $ref: "#/components/schemas/Session"
        - type: object
          required: [owner_token]
          properties:
            owner_token:
              type: string
              description: Returned only here. Keep it secret; it recovers host access.
...
    Room:
      type: object
      required: [id, name, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [active, closed]
        created_at:
          type: string
          format: date-time
...
    Member:
      type: object
      required: [id, display_name, role, joined_at, online]
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        joined_at:
          type: string
          format: date-time
        online:
          type: boolean
...